Fix: Update all migrations to use unsignedBigInteger for foreign keys
- Changed all integer mahasiswa_id, dosen_id foreign keys to unsignedBigInteger
- Fixed foreign key constraint errors in multiple migrations
- Successfully completed all pending migrations
- Tugas table and related tables now created successfully
- Fixed attendance_permits, gamification, kas_voting, academic tables
- All database tables now properly set up for dosen functionality

+++ database/migrations/2025_12_23_000103_create_attendance_logs_table.php
@@ -27,7 +27,7 @@ return new class extends Migration
             $table->foreignId('attendance_session_id')
                 ->constrained('attendance_sessions')
                 ->cascadeOnDelete();
-            $table->integer('mahasiswa_id');
+            $table->unsignedBigInteger('mahasiswa_id');
             $table->foreignId('attendance_token_id')
                 ->nullable()
                 ->constrained('attendance_tokens')
