+++ database/migrations/2026_01_14_100003_create_academic_tasks_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('academic_tasks', function (Blueprint $table) {
             $table->id();
-            $table->integer('mahasiswa_id');
+            $table->unsignedBigInteger('mahasiswa_id');
             $table->foreignId('mahasiswa_course_id')->constrained('mahasiswa_courses')->onDelete('cascade');
             $table->integer('meeting_number')->nullable();
             $table->string('title');
